Parse events and variables
Features: Parser can now read variable lists and event lists. It can read
the equations that define the events. It ignores comments too.
Lacking: Parser does not understand equations as functions but
rather simply as strings. Need to add an interpreter. None of the
parsed information initializes anything, so it does not interface
with the rest of the code yet.